pyproject.toml
[project]
name = "epstein-files-rag"
version = "0.1.0"
description = "Public RAG MVP for the Epstein Files using teyler/epstein-files-20k"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain",
    "langchain-community",
    "langchain-chroma",
    "langchain-openai",
    "chromadb",
    "sentence-transformers",
    "streamlit",
    "datasets",
    "google-generativeai",
    "python-dotenv"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
